disable low power mode, fix responsiveness
entering low power mode does not provide a meaningful savings
and was causing the responsiveness of the firmware to drop significantly
changing this fixed the boat's drive and will likely fix the antbots
as well.  Also moved from handling almost anything within
interrupts to setting flags picked up in the main loop.